Updating authentication middleware
